backend/requirements.txt
asgiref==3.8.1
Django==5.0.4
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
mysqlclient==2.2.4
pillow==10.3.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1

django-ckeditor~=6.7.1
python-slugify~=8.0.4